fix(frontend): address chat ds104 review feedback
