+++ packages/frontend/src/sections/RoomChat.tsx
@@ -19,6 +19,7 @@ import {
 import { copyToClipboard } from '../utils/clipboard';
 import { buildOpenHash } from '../utils/deepLink';
 import { toIsoFromLocalInput, toLocalDateTimeValue } from '../utils/datetime';
+import { resolveAttachmentKind } from '../utils/attachments';
 
 type ChatRoom = {
   id: string;
@@ -161,15 +162,6 @@ function transformLinkUri(uri?: string) {
 
 function sanitizeFilename(value: string) {
   return value.replace(/["\\\r\n]/g, '_').replace(/[/\\]/g, '_');
-}
-
-function resolveAttachmentKind(
-  mimeType: string | null | undefined,
-): AttachmentRecord['kind'] {
-  const normalized = (mimeType || '').toLowerCase();
-  if (normalized.startsWith('image/')) return 'image';
-  if (normalized === 'application/pdf') return 'pdf';
-  return 'file';
 }
 
 function toAttachmentRecord(attachment: {
@@ -2393,11 +2385,15 @@ export const RoomChat: React.FC = () => {
                         {canRevoke && (
                           <button
                             className="button secondary"
-                            onClick={() =>
+                            disabled={!!pendingUndoRevokeAck}
+                            onClick={() => {
+                              if (pendingUndoRevokeAck) {
+                                return;
+                              }
                               setPendingUndoRevokeAck({
                                 requestId: ackRequest.id,
-                              })
-                            }
+                              });
+                            }}
                           >
                             OK取消
                           </button>
