+++ packages/frontend/src/utils/attachments.ts
@@ -1,0 +1,10 @@
+export type AttachmentKind = 'image' | 'pdf' | 'file';
+
+export function resolveAttachmentKind(
+  mimeType: string | null | undefined,
+): AttachmentKind {
+  const normalized = (mimeType || '').toLowerCase();
+  if (normalized.startsWith('image/')) return 'image';
+  if (normalized === 'application/pdf') return 'pdf';
+  return 'file';
+}
